cache materials when importing
